Added WorkerTypes. Added offline validation code

+++ barrage_db.cpp
@@ -33,9 +33,6 @@ int main(int argc, char *argv[])
 	int duration;
 	std::string conf_file;
 	db_metadata db;
-	//std::vector<WorkerThread *> threads;
-	//WorkGroup group1;
-	std::cout << "Main: startup" << std::endl;
 
 	namespace po = boost::program_options;
 	po::options_description desc("barrageDB options");
@@ -46,7 +43,7 @@ int main(int argc, char *argv[])
 		("time", po::value<int>(&duration)->default_value(60), "set time to run (in seconds)")
 		("conf-file", po::value<std::string>(&conf_file)->default_value("confing.ini"), "configuration file name")
 		("validate-offline", "validate database contents offline")
-		("cleanup", po::value<int>(&cleanup)->default_value(1), "cleanup database on exit");
+		("cleanup", po::value<int>(&cleanup)->default_value(0), "cleanup database on exit");
 
 	po::variables_map var_map;
 	po::store(po::parse_command_line(argc, argv, desc), var_map);
@@ -59,33 +56,53 @@ int main(int argc, char *argv[])
 
 
 	try {
-		std::cout << "========= barrageDB ========" << std::endl;
-		std::cout << ">Number of threads  : " << thread_count << std::endl;
-		std::cout << ">Test duration      : " << duration << std::endl;
-		std::cout << ">configuration file : " << conf_file << std::endl;
-
 		db.load(conf_file);
-		db.print_contents();
+		//db.print_contents();
 
-		ret = check_table_exists(db.generate_connection_string());
-		if (!ret) {
-			create_table(db.generate_connection_string());
+		std::cout << "========= barrageDB ========" << std::endl;
+		if (var_map.count("validate-offline")) {
+			// Simply validate database contents and exit.
+
+			ret = check_table_exists(db.generate_connection_string());
+			if (ret == 0) {
+				std::cout << ">Validation mode: did not find barrage_data table in the database." << std::endl;
+				return 0;
+			} else if (ret == CONNECTION_ERROR) {
+				// raise exception
+			} 
+
+			// Could make this smart, have more worker threads to speed things up
+			WorkGroup validation_group(db.generate_connection_string(), 1, VALIDATOR);
+			validation_group.start();
+			validation_group.wait_all();
 		} else {
-			// Failure handling
+			std::cout << ">Number of threads  : " << thread_count << std::endl;
+			std::cout << ">Test duration      : " << duration << std::endl;
+			std::cout << ">configuration file : " << conf_file << std::endl;
+
+			ret = check_table_exists(db.generate_connection_string());
+			if (ret == 0) {
+				create_table(db.generate_connection_string());
+			} else if (ret == CONNECTION_ERROR) {
+				// raise exception
+			} else {
+				// Table exists, could validate before running?
+			}
+
+			WorkGroup group1(db.generate_connection_string(), thread_count, GENERATOR); 
+			group1.start();
+			group1.wait_all();
 		}
 
-		WorkGroup group1(db.generate_connection_string(), thread_count); 
-		group1.start();
-		group1.wait_all();
-
-		ret = drop_table(db.generate_connection_string());
+		if (cleanup) {
+			ret = drop_table(db.generate_connection_string());
+			// Check return type
+		}
 	}
 	catch (std::exception &e)
 	{
 		std::cout << "Error: " << e.what() << "\n";
 	}
-
-	std::cout << "Main: done" << std::endl;
 
 	return 0;
 }
